requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.6.1
pydantic-settings==2.1.0
openpyxl==3.1.2
pandas==2.2.0
python-multipart==0.0.9
python-dateutil==2.8.2
httpx==0.27.0
